Refactorisation of Pilots Logging. Database, service and client separated from PilotAgents
